All fixes till 11/06/2021

+++ FormsLoyalty/FormsLoyalty/ViewModels/CheckoutPageViewModel.cs
@@ -364,7 +364,7 @@ namespace FormsLoyalty.ViewModels
                 {
 
                     basketOrder.CouponsNo = SelectedCoupon.CouponID;
-                    totalTotal = AppData.Device.UserLoggedOnToDevice.Environment.Currency.FormatDecimal(basketOrder.TotalAmount - value.CouponValue);
+                    totalTotal = AppData.Device.UserLoggedOnToDevice.Environment.Currency.FormatDecimal(basketOrder.TotalAmount - (decimal)value.CouponValue);
                 }
                 else
                     totalTotal = AppData.Device.UserLoggedOnToDevice.Environment.Currency.FormatDecimal(basketOrder.TotalAmount);
